+++ package.json
@@ -8,8 +8,8 @@
     "start": "react-native start",
     "test": "jest",
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
-    "lint:fix": "eslint . --fix '{**/*,*}.{js,ts,jsx,tsx}'",
-    "format": "prettier --write '{**/*,*}.{js,ts,jsx,tsx}'",
+    "lint:fix": "eslint . --fix \"{**/*,*}.{js,ts,jsx,tsx}\"",
+    "format": "prettier --write \"{**/*,*}.{js,ts,jsx,tsx}\"",
     "prepare": "husky install"
   },
   "lint-staged": {
@@ -19,16 +19,14 @@
     ]
   },
   "dependencies": {
-    "@react-native-community/checkbox": "^0.5.12",
     "@react-navigation/native": "^6.0.10",
     "@react-navigation/native-stack": "^6.6.2",
+    "@rneui/theming": "^0.1.1",
     "formik": "^2.2.9",
     "react": "17.0.2",
     "react-native": "0.68.2",
-    "react-native-paper": "^4.12.1",
     "react-native-safe-area-context": "^4.3.1",
     "react-native-screens": "^3.13.1",
-    "react-native-vector-icons": "^9.1.0",
     "recoil": "^0.7.3",
     "styled-components": "^5.3.5",
     "yup": "^0.32.11"
