+++ src/assets/images/index.ts
@@ -1,4 +1,4 @@
-import CARPET_LOGO from './carpet_logo.png';
+import CARPET_LOGO from './carpet-logo.png';
 import CHECK from './check.png';
 
 export { CARPET_LOGO, CHECK };
